Refactor all classes into Duke package
