+++ src/data/duke.txt
@@ -1,1 +1,3 @@
-T | 0 | hello
+T | 0 | hello world
+E | 1 | sleep | 2020-12-12
+D | 0 | eat | 2022-02-02
