+++ src/main/java/duke/Command.java
@@ -1,3 +1,5 @@
+package duke;
+
 /**
  * Abstract class Command that represents commands on tasks.
  *
