+++ src/main/java/duke/ToDo.java
@@ -1,3 +1,5 @@
+package duke;
+
 /**
  * Class that encapsulates To Do task.
  *
