+++ src/main/java/duke/UnmarkCommand.java
@@ -1,23 +1,25 @@
+package duke;
+
 /**
- * Represents a Delete Command, to delete a task from tasks.
+ * Represents an Unmark Command, which unmarks a task as completed.
  *
  * @author Elgin
  */
-public class DeleteCommand extends Command {
+public class UnmarkCommand extends Command {
     /** The arguments of the command (e.g. 'sleep /at 2020-12-12'). */
     private final String arguments;
 
     /**
-     * Constructor of DeleteCommand.
+     * Constructor of MarkCommand.
      *
      * @param arguments The arguments of the command.
      */
-    public DeleteCommand(String arguments) {
+    public UnmarkCommand(String arguments) {
         this.arguments = arguments;
     }
 
     /**
-     * Executes the delete command and remove a task from tasks.
+     * Executes the Unmark command, and unmarks the task as not done.
      *
      * @param tasks All tasks present in Duke.
      * @param ui The UI controller that handles interaction between user and Duke.
@@ -25,8 +27,8 @@ public class DeleteCommand extends Command {
      */
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
-        String deletedTaskDesc = tasks.deleteItem(this.arguments);
+        String taskDescription = tasks.markOrUnmark(this.arguments, false);
 
-        ui.showTaskDeleted(tasks.getTaskLen(), deletedTaskDesc);
+        ui.showTaskUnmarked(taskDescription);
     }
 }
